Split up prelude-types.ts module
See the prior commit's message for rationale.
